==> 02_run_document.Rmd <==
---
title: "`r params$doc_title`"
output:
  flexdashboard::flex_dashboard:
    orientation: rows
runtime: shiny
css: style.css
params:
  doc_title: 2022 British Columbia Labour Market Outlook
resource_files:
- shiny_data/jo_500.csv
- shiny_data/occupation_outlook_table.csv
- shiny_data/occupation_outlook.csv
- shiny_data/industry_outlook.csv
- shiny_data/regional.csv
- shiny_data/ds_and_jo_tab.csv
- shiny_data/ds_and_jo.csv
- shiny_data/new_supply.csv
- raw_data/Definitions 2022E.xlsx
- R/functions.R
---

```{r}
#libraries---------------
library(tidyverse)
library(shinyWidgets)
library(plotly)
library(kableExtra)
library(wrapR) #  devtools::install_github("bcgov/wrapR")
library(leaflet)
library(DT)
library(here)
#constants------------
round_small <- 2 #round to 2 digits
round_medium <- -1 #round to nearest 10
round_large <- -3 #round to nearest 1000
#functions----------
source(here::here("R", "functions.R"))
#load data--------------
occupation_outlook <- vroom::vroom(here::here("shiny_data","occupation_outlook.csv"))
occupation_outlook_table <- vroom::vroom(here::here("shiny_data","occupation_outlook_table.csv"))
industry_outlook <- vroom::vroom(here::here("shiny_data","industry_outlook.csv"))
ds_and_jo_tab <- vroom::vroom(here::here("shiny_data","ds_and_jo_tab.csv"))
ds_and_jo <-  vroom::vroom(here::here("shiny_data","ds_and_jo.csv"))
regional <-  vroom::vroom(here::here("shiny_data","regional.csv"))
new_supply <- vroom::vroom(here::here("shiny_data","new_supply.csv"))
jo_500 <-  vroom::vroom(here::here("shiny_data","jo_500.csv"))
definitions <- readxl::read_excel(here::here("raw_data","Definitions 2022E.xlsx"))
hoo <- jo_500%>%
  filter(hoo!=FALSE)%>%
  select(-hoo)
jo_500 <- jo_500%>%
  select(-hoo)

current_year <- min(ds_and_jo$date)-1
page_titles <- c("Landing Page",
                 "Highlights Part 1",
                 "Annual Outlook",
                 "Regional Outlook",
                 "Industry Outlook",
                 "Occupational Outlook",
                 "Industry Outlook Data Table",
                 "Occupational Outlook Data Table",
                 "High Opportunity Occupations",
                 "Job Openings 500 Occupations",
                 "Sources of New Supply",
                 "Labour Market Definitions")
make_title <- function(strng){
  strng <- str_to_title(str_replace_all(strng,"_"," "))
}

nav_but <- function(direction, current_page){
    if(direction=="previous"){
      if(current_page==1){
        name = tail(page_titles, n=1)
      }else{
        name = page_titles[current_page-1]
      }
    }
  if(direction == "next"){
      if(current_page == length(page_titles)){
        name = page_titles[1]
      }else{
        name = page_titles[current_page+1]
      }
    }
  name <- str_to_lower(name)%>%
      str_replace_all(" ","-")
  lab <-  str_to_title(paste(direction, "Page"))
  oc <- paste0("location.href='#section-",name,"';")
  actionButton("1", label = lab, class = "btn btn-sm btn-primary", 
            onclick = oc)
}

```

`r page_titles[1]` {data-navmenu="Table of Contents"}
=====================================
Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 1)
nav_but(direction = "next", current_page = 1)
```

Row {data-height=600}
-------------------------------------

```{r, out.width = "125%"}
knitr::include_graphics("aest.png")#the cover image
```

Row {data-height=400}
-------------------------------------

###

<h1 class="center">`r params$doc_title`.</h1> <!-- add the title -->

<br><br>

<h3> The B.C. Labour Market Outlook is a 10-year forecast of the expected flow of supply and demand for labour in the province. It is revised every year with the most up-to-date information available. The purpose is to provide British Columbians with the knowledge required to make informed decisions on careers, hiring, training and education.</h3>



`r page_titles[2]`  {data-navmenu="Table of Contents"}
===========================================

### `r page_titles[2]` | `r paste(current_year, current_year+10, sep="-")`

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 2)
nav_but(direction = "next", current_page = 2)

```


Row {data-height=1000}
-------------------------------------
   
###     
<br>
```{r}
tbbl <- ds_and_jo_tab %>%
  kbl(format.args = list(big.mark = ",")) %>%
  kable_styling()%>%
  row_spec(c(1,4), bold = TRUE, background = "#e0e0e0")
gsub("<thead>.*</thead>", "", tbbl)
```   

### Job Openings by Education Level   

```{r, fig.retina=2}
jo_by_ed <- read_csv(here("shiny_data","jo_by_ed.csv"))
ct <- paste(scales::comma(round(sum(jo_by_ed$`Job Openings`)+1, round_large)),"Job Openings \n Expected in B.C")
wrapR::donut(jo_by_ed, `Job Openings`, group_var = Typical_Education, center_text = ct, center_text_size = 2.5, lab_size=2.5)
```

`r page_titles[3]`  {data-navmenu="Table of Contents"}
===========================================

### `r page_titles[3]` | `r paste(current_year, current_year+10, sep="-")`

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 3)
nav_but(direction = "next", current_page = 3)
```

### Inputs

```{r}
selectInput(
  "annual_outlook",
  "Select a Region",
  unique(ds_and_jo$geographic_area),
  selected = unique(ds_and_jo$geographic_area)[1],
  multiple = FALSE)

col_data <- reactive({
  filter(ds_and_jo,
       geographic_area == input$annual_outlook,
       name %in% c("Expansion Demand", "Replacement Demand"))
})
pie_data <- reactive({
  col_data()%>%
  group_by(name)%>%
  summarise(value=round(sum(value), round_large))
})
ct_reactive <- reactive({
  ct <- paste(scales::comma(round(sum(pie_data()$value)+1, round_large)),"Job Openings \n Expected in",input$annual_outlook)
})

```

Row {data-height=500}
-------------------------------------
###
```{r, fig.retina=2}
renderPlot({
  wrapR::donut(pie_data(), value , group_var = name, center_text = ct_reactive(), center_text_size = 4, lab_size=4)
})
```
###
```{r}
renderPlot({
  ggplot(col_data(), aes(factor(date),value, fill=name))+
  geom_col()+
  theme_minimal()+
  scale_fill_viridis_d()+
  labs(x="",y="")+ 
  theme(text = element_text(size = 14))  
})
```
Row {data-height=500}
-------------------------------------
###

```{r, fig.retina=2}
pie_dat <- ds_and_jo_tab%>%
  filter(name %in% c(
    "Young People Starting Work",
    "Immigrants",
    "Migrants From Other Provinces",
    "Additional Supply Requirement",
    "Decline Unemployment"))
wrapR::donut(pie_dat, value , group_var = name, center_text = ct, center_text_size = 4, lab_size=4)
```

###

```{r, fig.retina=2}
bar_dat <- ds_and_jo%>%
  filter(geographic_area=="British Columbia",
         name %in% c(
           "Young People Starting Work",
           "Immigrants",
           "Migrants From Other Provinces",
           "Additional Supply Requirement",
           "Decline Unemployment",
           "Labour Force Exits"
         ))%>%
  mutate(date=factor(date))

plt <- wrapR::parts_and_whole(bar_dat, date, value, name, "stack")+
  theme_minimal()+
  labs(x="", y="")
  
myplot <- plotly::ggplotly(plt)%>%
    config(displayModeBar = F)
#plotly messes up the legend... fix it
for (i in 1:length(myplot$x$data)){
  if (!is.null(myplot$x$data[[i]]$name)){
    myplot$x$data[[i]]$name = gsub('^\\(|,\\d+\\)$', '', myplot$x$data[[i]]$name)
  }
}
myplot
```

`r page_titles[4]`  {data-navmenu="Table of Contents"}
=========================================

### `r page_titles[4]`

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 4)
nav_but(direction = "next", current_page = 4)
```

### Inputs

```{r}
plot_vars <- make_title(c(str_subset(unique(regional$name),"per_capita"),"employment_growth"))
geo_vars <- make_title(unique(regional$geographic_area))
table_vars <- str_subset(unique(regional$name),"per_capita", negate = TRUE)
selectInput(
  "regional_map",
  "Select a Variable for Map",
  plot_vars,
  selected = plot_vars[1],
  multiple = FALSE)

selectInput(
  "regional_table",
  "Select a Region for Table",
  geo_vars,
  selected = geo_vars[1],
  multiple = FALSE)

map_tbbl <- reactive({
  regional%>%
    mutate(name=make_title(name))%>%#do this so it matches with filter
    filter(name==input$regional_map)
})

table_tbbl <- reactive({
  regional%>%
    mutate(geographic_area=make_title(geographic_area))%>%
    filter(geographic_area==input$regional_table,
           name %in% table_vars)%>%
    camel_to_title()
})
```

Row 
-------------------------------------
###
```{r}
leaflet::renderLeaflet({
  bc_reg_choro(map_tbbl(), geographic_area, name, value, "percent")
})
```

###

<h1 class="center">
```{r} 
renderUI({table_tbbl()$geographic_area[1]}) 
```
</h1>

<br>

<table border frame="box" class="myFormat">
  <tr>
  <td>
  ```{r} 
    renderUI({table_tbbl()$name[1]})
  ```
  </td>
  <td>
  ```{r} 
  renderUI({table_tbbl()$name[2]})
  ```
  </td>
  <td>
  ```{r}
  renderUI({table_tbbl()$name[3]})
  ```
  </td>
  </tr>
  <tr>
  <td>
  ```{r}
  renderUI({scales::comma(table_tbbl()$value[1], accuracy=1000)})
  ```
  </td>
  <td>
  ```{r}
  renderUI({scales::percent(table_tbbl()$value[2], accuracy=.1)})
  ```
  </td>
  <td>
  ```{r}
  renderUI({scales::comma(table_tbbl()$value[3], accuracy=1000)})
  ```
  </td>
  </tr>
  </table>

<br>

<table border frame="box" class="myFormat">
  <tr>
  <td>
  ```{r} 
  renderUI({table_tbbl()$name[4]})
  ```
  </td>
  <td>
  ```{r}
  renderUI({table_tbbl()$name[5]})
  ```
  </td>
  </tr>
  <tr>
  <td>
  ```{r}
  renderUI({scales::comma(table_tbbl()$value[4], accuracy=1000)})
  ```
  </td>
  <td>
  ```{r}
  renderUI({scales::comma(table_tbbl()$value[5], accuracy=1000)})
  ```
  </td>
  </tr>
  </table>



`r page_titles[5]` {data-navmenu="Table of Contents"}
=========================================

### `r page_titles[5]`

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 5)
nav_but(direction = "next", current_page = 5)

```
### Inputs

```{r}
selectizeGroupUI(
        id = "industry_filters",
        inline = FALSE,
        params = list(
          var_one = list(inputId = "name", title = "Choose a Variable", placeholder = "select"),
          var_two = list(inputId = "Geographic_Area", title = "Choose a Geographic Area", placeholder = "select"),
          var_three = list(inputId = "aggregate_industry", title = "Choose an Aggregate Industry", placeholder = "select")
        )
      )

industry_outlook_plot <- industry_outlook%>%
  filter(!name %in% c("First Five Cagr","Second Five Cagr", "Annual Replacement Rate"))%>%
  pivot_wider()%>%
  rename("Job Openings {current_year+1}-{current_year+10}" := `Job Openings`,
        "Employment {current_year}" := `Employment Level`,
        "Employment Growth {current_year+1}-{current_year+10}" := `Ten Year Cagr`,
        "Expansion Demand {current_year+1}-{current_year+10}" := `Expansion Demand`,
        "Replacement Demand {current_year+1}-{current_year+10}" := `Replacement Demand`)%>%
  pivot_longer(cols=-c(Geographic_Area, aggregate_industry, industry)) #if want to reorder facets could do it here: mutate(name=factor(name, levels=....))

industry_mod <- callModule(
    module = selectizeGroupServer,
    id = "industry_filters",
    data = industry_outlook_plot,
    vars = c(
      "name",
      "Geographic_Area",
      "aggregate_industry"
    )
)

industry_agg <- reactive({
  case_when(is.null(input[["industry_filters-aggregate_industry"]]) ~ "aggregate_industry",
            TRUE~"industry")
})

industry_tbbl <- reactive({
  industry_mod()%>%
    distinct(Geographic_Area, name, get(industry_agg()), .keep_all = TRUE)%>%
    select(Geographic_Area, name, value, industry ="get(industry_agg())")%>%
    na.omit()
})

```

Row 
-------------------------------------
###

```{r fig.retina=2}
renderPlotly({
  plt <- industry_tbbl()%>%
    ggplot(aes(value,
               fct_reorder(industry, value, .fun = max),
               fill= fct_reorder(Geographic_Area, value, .fun = max),
               text= paste0(
                "Geographic Area: ", Geographic_Area, "\n",
                "Industry: ", industry, "\n",
                "Value: ", scales::comma(value, accuracy = .01))
               ))+
    geom_col(position = "dodge2")+
    scale_x_continuous(labels = scales::comma)+
    scale_fill_viridis_d()+
    facet_wrap(vars(name), scales = "free_x")+
    labs(x="",y="")+
    guides(fill = guide_legend(reverse=TRUE))+
    theme_minimal()+
    theme(axis.text.x = element_text(angle = 15))
  ggplotly(plt, tooltip="text")%>%
    config(displayModeBar = F)
})
```

`r page_titles[6]`  {data-navmenu="Table of Contents"}
========================================

### `r page_titles[6]` 

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 6)
nav_but(direction = "next", current_page = 6)
```

### Inputs

```{r}
selectizeGroupUI(
        id = "occupation_filters",
        inline = FALSE,
        params = list(
          var_one = list(inputId = "name", title = "Choose a Variable", placeholder = "select"),
          var_two = list(inputId = "Geographic_Area", title = "Choose a Geographic Area", placeholder = "select"),
          var_three = list(inputId = "NOC1", title = "Choose a NOC1", placeholder = "select"),
          var_four = list(inputId = "NOC2", title = "Choose a NOC2", placeholder = "select"),
          var_five = list(inputId = "NOC3", title = "Choose a NOC3", placeholder = "select")
        )
      )

occupation_outlook_plot <- occupation_outlook%>%
  filter(!name %in% c("First Five Cagr","Second Five Cagr", "Annual Replacement Rate"))%>%
  select(-Typical_Education)%>%
  pivot_wider()%>%
  rename("Job Openings {current_year+1}-{current_year+10}" := `Job Openings`,
        "Employment {current_year}" := `Employment Level`,
        "Employment Growth {current_year+1}-{current_year+10}" := `Ten Year Cagr`,
        "Expansion Demand {current_year+1}-{current_year+10}" := `Expansion Demand`,
        "Replacement Demand {current_year+1}-{current_year+10}" := `Replacement Demand`)%>%
  pivot_longer(cols=-c(Geographic_Area, NOC1, NOC2, NOC3, NOC4))

occupation_mod <- callModule(
    module = selectizeGroupServer,
    id = "occupation_filters",
    data = occupation_outlook_plot,
    vars = c(
      "name",
      "Geographic_Area",
      "NOC1",
      "NOC2",
      "NOC3"
    )
)

noc_agg <- reactive({
  case_when(is.null(input[["occupation_filters-NOC1"]]) & 
           is.null(input[["occupation_filters-NOC2"]]) &
           is.null(input[["occupation_filters-NOC3"]]) ~ "NOC1",
           is.null(input[["occupation_filters-NOC2"]]) & 
           is.null(input[["occupation_filters-NOC3"]]) ~ "NOC2",
           is.null(input[["occupation_filters-NOC3"]]) ~ "NOC3",
           TRUE~"NOC4")
})

occupation_tbbl <- reactive({
  occupation_mod()%>%
    distinct(Geographic_Area, name, get(noc_agg()), .keep_all = TRUE)%>%
    select(Geographic_Area, name, value, noc="get(noc_agg())")%>%
    na.omit()
})

```

Row 
-------------------------------------------

###

```{r fig.retina=2}
renderPlotly({
  plt <- occupation_tbbl()%>%
    ggplot(aes(value,
               fct_reorder(noc, value, .fun = max),
               fill= fct_reorder(Geographic_Area, value, .fun = max),
               text= paste0(
                "Geographic Area: ", Geographic_Area, "\n",
                "NOC: ", noc, "\n",
                "Value: ", scales::comma(value, accuracy = .01))
               ))+
    geom_col(position = "dodge2")+
    scale_x_continuous(labels = scales::comma)+
    scale_fill_viridis_d()+
    facet_wrap(vars(name), scales = "free_x")+
    labs(x="",y="")+
    guides(fill = guide_legend(reverse=TRUE))+
    theme_minimal()+
    theme(axis.text.x = element_text(angle = 15))
  ggplotly(plt, tooltip="text")%>%
    config(displayModeBar = F)
})
```

`r page_titles[7]`  {data-navmenu="Table of Contents"}
========================================

### `r page_titles[7]`

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 7)
nav_but(direction = "next", current_page = 7)
```

### Inputs

```{r}
selectizeGroupUI(
        id = "industry_filters_table",
        inline = FALSE,
        params = list(
         # var_one = list(inputId = "name", title = "Choose a Variable", placeholder = "select"),
          var_two = list(inputId = "Geographic_Area", title = "Choose a Geographic Area", placeholder = "select"),
          var_three = list(inputId = "aggregate_industry", title = "Choose an Aggregate Industry", placeholder = "select")
        )
      )

industry_outlook_table <- industry_outlook%>%
  pivot_wider()%>%
  rename("Annual Replacement Rate(%)"=`Annual Replacement Rate`,
         "Employment {current_year}" := `Employment Level`,
         "Job Openings As Share Of Employment(%)" = `Job Openings As Share Of Employment`,
         "Employment Growth(%) {current_year+1}-{current_year+5}" := `First Five Cagr`,
         "Employment Growth(%) {current_year+1}-{current_year+10}" := `Ten Year Cagr`,
         "Employment Growth(%) {current_year+6}-{current_year+10}" := `Second Five Cagr`,
         "Job Openings Expansion {current_year+1}-{current_year+10}" := `Expansion Demand`,
         "Job Openings Replacement {current_year+1}-{current_year+10}" := `Replacement Demand`,
         "Job Openings {current_year+1}-{current_year+10}" := `Job Openings`)%>%
  pivot_longer(cols=-c(Geographic_Area, aggregate_industry, industry)) 

industry_mod_table <- callModule(
    module = selectizeGroupServer,
    id = "industry_filters_table",
    data = industry_outlook_table,
    vars = c(
   #   "name",
      "Geographic_Area",
      "aggregate_industry"
    )
)

industry_agg_table <- reactive({
  case_when(is.null(input[["industry_filters_table-aggregate_industry"]]) ~ "aggregate_industry",
            TRUE~"industry")
})

industry_tbbl_table <- reactive({
  industry_mod_table()%>%
    distinct(Geographic_Area, name, get(industry_agg_table()), .keep_all = TRUE)%>%
    select(Geographic_Area, name, value, industry ="get(industry_agg_table())")%>%
#    na.omit()%>%
    pivot_wider()
})

```


Row
-------------------------------------
### 

```{r}
DT::renderDT(
  industry_tbbl_table()%>%
    fix_col_names()%>%
    datatable(rownames=FALSE)%>%
    formatStyle(
  'Industry',
  target = 'row',
  backgroundColor = styleEqual(c(NA), c('lightgray')))
)
```


`r page_titles[8]` {data-navmenu="Table of Contents"}
========================================

### `r page_titles[8]`

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 8)
nav_but(direction = "next", current_page = 8)

```

### Inputs

```{r}
selectizeGroupUI(
        id = "occupation_filters_table",
        inline = FALSE,
        params = list(
          var_one = list(inputId = "Typical_Education", title = "Choose an education level", placeholder = "select"),
          var_two = list(inputId = "Geographic_Area", title = "Choose a Geographic Area", placeholder = "select"),
          var_three = list(inputId = "NOC1", title = "Choose a NOC1", placeholder = "select"),
          var_four = list(inputId = "NOC2", title = "Choose a NOC2", placeholder = "select"),
          var_five = list(inputId = "NOC3", title = "Choose a NOC3", placeholder = "select")
        )
      )

occupation_outlook_table <- occupation_outlook_table%>%
  select(-noc)%>%
  filter(name %in% c("Employment Level","Expansion Demand","Replacement Demand","Job Openings", "Job Openings As Share Of Employment"))%>%
  pivot_wider()%>%
  rename("Job Openings {current_year+1}-{current_year+10}" := `Job Openings`,
         "Job Openings As Share Of Employment(%)" = `Job Openings As Share Of Employment`,
        "Employment {current_year}" := `Employment Level`,
        "Expansion Demand {current_year+1}-{current_year+10}" := `Expansion Demand`,
        "Replacement Demand {current_year+1}-{current_year+10}" := `Replacement Demand`)%>%
  pivot_longer(cols=-c(Geographic_Area, Typical_Education, NOC1, NOC2, NOC3, NOC4))

occupation_mod_table <- callModule(
    module = selectizeGroupServer,
    id = "occupation_filters_table",
    data = occupation_outlook_table,
    vars = c(
      "Typical_Education",
      "Geographic_Area",
      "NOC1",
      "NOC2",
      "NOC3"
    )
)

noc_agg_table <- reactive({
  case_when(is.null(input[["occupation_filters_table-NOC1"]]) &
           is.null(input[["occupation_filters_table-NOC2"]]) &
           is.null(input[["occupation_filters_table-NOC3"]]) ~ "NOC1",
           is.null(input[["occupation_filters_table-NOC2"]]) &
           is.null(input[["occupation_filters_table-NOC3"]]) ~ "NOC2",
           is.null(input[["occupation_filters_table-NOC3"]]) ~ "NOC3",
           TRUE~"NOC4")
})

occupation_tbbl_table <- reactive({
  occupation_mod_table()%>%
    distinct(Geographic_Area, name, get(noc_agg_table()), .keep_all = TRUE)%>%
    select(Geographic_Area, Typical_Education, name, value, noc="get(noc_agg_table())")%>%
    pivot_wider()
})

```


Row 
-------------------------------------
### 

```{r}
DT::renderDT(
  occupation_tbbl_table()%>%
    fix_col_names()%>%
    datatable(rownames=FALSE)%>%
    formatStyle(
  'NOC',
  target = 'row',
  backgroundColor = styleEqual(c(NA), c('lightgray')))
)
```

`r page_titles[9]` {data-navmenu="Table of Contents"}
========================================

### `r page_titles[9]`


Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 9)
nav_but(direction = "next", current_page = 9)
```

### Inputs


```{r}
selectizeGroupUI(
        id = "hoo_filters",
        inline = FALSE,
        params = list(
          var_one = list(inputId = "Typical_Education", title = "Choose an education level", placeholder = "select"),
          var_two = list(inputId = "Geographic_Area", title = "Choose a Geographic Area", placeholder = "select"),
          var_three = list(inputId = "NOC1", title = "Choose a NOC1", placeholder = "select"),
          var_four = list(inputId = "NOC2", title = "Choose a NOC2", placeholder = "select"),
          var_five = list(inputId = "NOC3", title = "Choose a NOC3", placeholder = "select"),
          var_six = list(inputId = "interests", title = "Choose your Interests", placeholder = "select")
        )
      )


hoo_mod <- callModule(
    module = selectizeGroupServer,
    id = "hoo_filters",
    data = hoo,
    vars = c(
      "Typical_Education",
      "Geographic_Area",
      "NOC1",
      "NOC2",
      "NOC3",
      "interests"
    )
)

hoo_table <- reactive({
  hoo_mod()%>%
    distinct(Geographic_Area, Typical_Education, name, NOC4, interests, .keep_all = TRUE)%>%
    select(noc=NOC4, everything())%>%
    select(-NOC1,-NOC2,-NOC3)%>%
    pivot_wider()
})

```


Row 
-------------------------------------
### 

```{r}

DT::renderDT(
  hoo_table()%>%
    fix_col_names()%>%
    datatable(rownames=FALSE)
)
```






`r page_titles[10]`  {data-navmenu="Table of Contents"}
========================================

### `r page_titles[10]`

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 10)
nav_but(direction = "next", current_page = 10)
```

### Inputs


```{r}

selectizeGroupUI(
        id = "jo_500_filters",
        inline = FALSE,
        params = list(
          var_one = list(inputId = "Typical_Education", title = "Choose an education level", placeholder = "select"),
          var_two = list(inputId = "Geographic_Area", title = "Choose a Geographic Area", placeholder = "select"),
          var_three = list(inputId = "NOC1", title = "Choose a NOC1", placeholder = "select"),
          var_four = list(inputId = "NOC2", title = "Choose a NOC2", placeholder = "select"),
          var_five = list(inputId = "NOC3", title = "Choose a NOC3", placeholder = "select"),
          var_six = list(inputId = "interests", title = "Choose your Interests", placeholder = "select")
        )
      )

jo_500_mod <- callModule(
    module = selectizeGroupServer,
    id = "jo_500_filters",
    data = jo_500,
    vars = c(
      "Typical_Education",
      "Geographic_Area",
      "NOC1",
      "NOC2",
      "NOC3",
      "interests"
    )
)

jo_500_table <- reactive({
  jo_500_mod()%>%
    distinct(Geographic_Area, name, NOC4, interests, .keep_all = TRUE)%>%
    select(noc=NOC4, everything())%>%
    select(-NOC1,-NOC2,-NOC3)%>%
    pivot_wider()%>%
    na.omit()
})

```


Row 
-------------------------------------
### 

```{r}

DT::renderDT(
  jo_500_table()%>%
    fix_col_names()%>%
    datatable(rownames=FALSE)
)
```

`r page_titles[11]`  {data-navmenu="Table of Contents"} 
=========================================

### `r page_titles[11]` 

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 11)
nav_but(direction = "next", current_page = 11)
```


### Inputs


```{r}

selectizeGroupUI(
        id = "new_supply_filters",
        inline = FALSE,
        params = list(
          var_one = list(inputId = "NOC1", title = "Choose a NOC1", placeholder = "select"),
          var_two = list(inputId = "NOC2", title = "Choose a NOC2", placeholder = "select"),
          var_three = list(inputId = "NOC3", title = "Choose a NOC3", placeholder = "select")
         )
      )

new_supply_mod <- callModule(
    module = selectizeGroupServer,
    id = "new_supply_filters",
    data = new_supply,
    vars = c(
      "NOC1",
      "NOC2",
      "NOC3"
   )
)

new_supply_table <- reactive({
  new_supply_mod()%>%
    select(noc=NOC4, everything())%>%
    select(-NOC1,-NOC2,-NOC3)
})

```


Row 
-------------------------------------
### 

```{r}

DT::renderDT(
  new_supply_table()%>%
    fix_col_names()%>%
    datatable(rownames=FALSE)
)
```






`r page_titles[12]`  {data-navmenu="Table of Contents"}
=========================================

### `r page_titles[12]` 

Inputs {.sidebar}
-------------------------------------

### Navigation

```{r}
nav_but(direction = "previous", current_page = 12)
nav_but(direction = "next", current_page = 12)
```

Row 
-------------------------------------
###

```{r}
DT::datatable(definitions, options = list(dom="t", paging=FALSE), rownames = FALSE)
```

###

```{r}
knitr::include_graphics("workers.jpg")
```
